+++ src/components/Generator/Generator.js
@@ -1,5 +1,5 @@
 import FlourSelection from "./FlourSelection";
-import StandardButton from "../StandardButton";
+
 import "./FlourSelection.css";
 
 export default function Generator() {
