+++ src/components/RoutingButton.js
@@ -1,6 +1,6 @@
-import "./StandardButton.css";
+import "./RoutingButton.css";
 import { useHistory } from "react-router-dom";
-export default function StandardButton({ name, route, look }) {
+export default function RoutingButton({ name, route, look }) {
   let history = useHistory();
   function handleClick() {
     history.push(`${route}`);
